introduce VRFs
this adds support for VRFs (Virtual Routing and Forwarding
instances) through libvrf from vrf-tools. pptpd gains the ability
to have its TCP and GRE sockets in a VRF different from the one
the process is running in.

As pppd is started from pptpd and does not have VRF support, the
pppX devices will be in the initial VRF. The PPTP traffic will
however be flowing in the given VRF. This allows full separation
of encapsulated and encapsulating networks.

Signed-off-by: David Lamparter <david.lamparter@adyton.net>
Cc: David Lamparter <equinox@diac24.net>
Cc: Franco Fichtner <franco.fichtner@adyton.net>
Acked-by: James Cameron <quozl@laptop.org>